Code cleaning

+++ scan.go
@@ -201,32 +201,32 @@ func parseRDB(path string) []Game {
 					pos += len
 				}
 			case "size":
-				pow := int(rdb[pos]) - 0xC9 // CC -> 8, CD -> 16, CE -> 32, CD -> 64
-				len := int(math.Pow(float64(2), float64(pow))) / 8
+				pow := float64(rdb[pos]) - 0xC9 // CC -> 8, CD -> 16, CE -> 32, CD -> 64
+				len := int(math.Pow(2, pow)) / 8
 				fmt.Println(len)
 				pos++
 				pos += len
 			case "releaseyear":
-				pow := int(rdb[pos]) - 0xC9 // CC -> 8, CD -> 16, CE -> 32, CD -> 64
-				len := int(math.Pow(float64(2), float64(pow))) / 8
+				pow := float64(rdb[pos]) - 0xC9 // CC -> 8, CD -> 16, CE -> 32, CD -> 64
+				len := int(math.Pow(2, pow)) / 8
 				fmt.Println(len)
 				pos++
 				pos += len
 			case "releasemonth":
-				pow := int(rdb[pos]) - 0xC9 // CC -> 8, CD -> 16, CE -> 32, CD -> 64
-				len := int(math.Pow(float64(2), float64(pow))) / 8
+				pow := float64(rdb[pos]) - 0xC9 // CC -> 8, CD -> 16, CE -> 32, CD -> 64
+				len := int(math.Pow(2, pow)) / 8
 				fmt.Println(len)
 				pos++
 				pos += len
 			case "users":
-				pow := int(rdb[pos]) - 0xC9 // CC -> 8, CD -> 16, CE -> 32, CD -> 64
-				len := int(math.Pow(float64(2), float64(pow))) / 8
+				pow := float64(rdb[pos]) - 0xC9 // CC -> 8, CD -> 16, CE -> 32, CD -> 64
+				len := int(math.Pow(2, pow)) / 8
 				fmt.Println(len)
 				pos++
 				pos += len
 			case "coop":
-				pow := int(rdb[pos]) - 0xC9 // CC -> 8, CD -> 16, CE -> 32, CD -> 64
-				len := int(math.Pow(float64(2), float64(pow))) / 8
+				pow := float64(rdb[pos]) - 0xC9 // CC -> 8, CD -> 16, CE -> 32, CD -> 64
+				len := int(math.Pow(2, pow)) / 8
 				fmt.Println(len)
 				pos++
 				pos += len
